Updated my skills section, started ignoring logs
Added haskell, as I've started messing around with it.

Clarified my experience with python, I started using it to supplement the lab
reports I was writing for my Physics courses. That was done with numpy, but I
don't feel it's important to mention it as a specific library like I did with
django.

Fixed some formatting issues, where python was not capitalized like the other
skill headers.

I don't really feel the need to track changes made to logfiles.